minor track performance enhancements
- Removed some unnecessary variables.
- Removed some unnecessary castings.
- Delayed some checks.
- Clamped another variable to prevent crashing.

+++ src/main/java/train/client/render/models/blocks/ModelMediumStraightTCTrack.java
@@ -35,7 +35,7 @@ public class ModelMediumStraightTCTrack extends ModelBase {
 		FMLClientHandler.instance().getClient().renderEngine.bindTexture(new ResourceLocation(Info.resourceLocation, Info.modelTexPrefix + "track_normal.png"));
 		GL11.glColor3f(1, 1, 1);
 		//GL11.glScalef(0.5f, 0.5f, 0.5f);
-		int facing = tcRail.getWorldObj().getBlockMetadata((int) tcRail.xCoord, (int) tcRail.yCoord, (int) tcRail.zCoord);
+		int facing = tcRail.getWorldObj().getBlockMetadata(tcRail.xCoord, tcRail.yCoord, tcRail.zCoord);
 
 		if (facing == 3) {
 			GL11.glRotatef(-90, 0, 1, 0);
